#0: add height sharding support for rm transpose hw
