+++ ttnn/cpp/ttnn/operations/data_movement/transpose/device/transpose_op.cpp
@@ -153,7 +153,11 @@ operation::ProgramWithCallbacks Transpose::create_program(const std::vector<Tens
     switch (parallelization_strategy) {
         case TransposeOpParallelizationStrategy::MULTI_CORE_WH:
             if (input_tensor.is_sharded()) {
-                return detail::transpose_wh_multi_core_sharded(input_tensor, output_tensor);
+                if (input_tensor.get_layout() == Layout::ROW_MAJOR) {
+                    return detail::transpose_wh_multi_core_sharded_rm(input_tensor, output_tensor);
+                } else {
+                    return detail::transpose_wh_multi_core_sharded(input_tensor, output_tensor);
+                }
             } else {
                 return detail::transpose_wh_multi_core(input_tensor, output_tensor);
             }
