+++ ttnn/cpp/ttnn/operations/data_movement/transpose/device/transpose_op.hpp
@@ -33,10 +33,4 @@ struct Transpose {
 };
 
 
-operation::ProgramWithCallbacks transpose_wh_multi_core(const Tensor &a, Tensor &output);
-operation::ProgramWithCallbacks transpose_wh_multi_core_sharded(const Tensor &a, Tensor &output);
-operation::ProgramWithCallbacks transpose_hc_multi_core(const Tensor &a, Tensor &output);
-operation::ProgramWithCallbacks transpose_hc_multi_core_sharded(const Tensor &a, Tensor &output);
-operation::ProgramWithCallbacks transpose_cn_multi_core(const Tensor &a, Tensor &output);
-
 }
